cleaned up service classes
