+++ pubspec.yaml
@@ -53,4 +53,3 @@ flutter:
   uses-material-design: true
   assets:
     - assets/welcomescreen.png
-
